Move Base splitter from main app models to order management engine services
