+++ engines/order_management/app/services/order_management/stock/basic_splitter.rb
@@ -1,0 +1,29 @@
+# frozen_string_literal: true
+
+module OrderManagement
+  module Stock
+    class BasicSplitter
+      attr_reader :packer, :next_splitter
+
+      def initialize(packer, next_splitter = nil)
+        @packer = packer
+        @next_splitter = next_splitter
+      end
+      delegate :stock_location, :order, to: :packer
+
+      def split(packages)
+        return_next(packages)
+      end
+
+      private
+
+      def return_next(packages)
+        next_splitter ? next_splitter.split(packages) : packages
+      end
+
+      def build_package(contents = [])
+        @packer.package_factory.new(stock_location, order, contents)
+      end
+    end
+  end
+end
