+++ spec/config/application_spec.rb
@@ -3,7 +3,7 @@ require 'spec_helper'
 describe Openfoodnetwork::Application, 'configuration' do
   let(:config) { described_class.config }
 
-  it "sets Spree::Stock::Splitter::Base as the only stock splitter" do
-    expect(config.spree.stock_splitters).to eq [Spree::Stock::Splitter::Base]
+  it "sets OrderManagement::Stock::BasicSplitter as the only stock splitter" do
+    expect(config.spree.stock_splitters).to eq [OrderManagement::Stock::BasicSplitter]
   end
 end
